rename changelog file to avoid it being missing
